requirements.txt
asgiref==3.2.3
astropy==3.2.3
astroquery==0.3.10
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
Django==3.0
django-filter==2.2.0
djangorestframework==3.10.3
html5lib==1.0.1
idna==2.8
importlib-metadata==1.0.0
jeepney==0.4.1
keyring==19.3.0
kiwisolver==1.1.0
matplotlib==3.1.2
more-itertools==8.0.0
msgpack==0.6.2
msgpack-numpy==0.4.4.3
numpy==1.17.4
pandas==0.25.3
psycopg2-binary==2.8.4
pycparser==2.19
pyparsing==2.4.5
PySide2==5.13.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
SecretStorage==3.1.1
shiboken2==5.13.2
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
webencodings==0.5.1
zipp==0.6.0
